Moved firewalld to its own namespace.
