+++ library/network/src/lib/y2firewall/firewalld.rb
@@ -1,0 +1,75 @@
+# encoding: utf-8
+#
+# ***************************************************************************
+#
+# Copyright (c) 2017 SUSE LLC.
+# All Rights Reserved.
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of version 2 or 3 of the GNU General
+# Public License as published by the Free Software Foundation.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.   See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, contact SUSE LLC.
+#
+# To contact SUSE about this file by physical or electronic mail,
+# you may find current contact information at www.suse.com
+#
+# ***************************************************************************
+
+require "y2firewall/firewalld/api"
+
+module Y2Firewall
+  # Main class to interact with Firewalld
+  class Firewalld
+    extend Forwardable
+    attr_accessor :api
+
+    PACKAGE = "firewalld".freeze
+    SERVICE = "firewalld".freeze
+
+    def_delegators :@api, :enable!, :disable!
+
+    def initialize
+      @api = Y2Firewall::Firewalld::Api.new
+    end
+
+    def running?
+      api.running?
+    end
+
+    def installed?
+      return true if @installed
+
+      @installed = PackageSystem.Installed(PACKAGE)
+    end
+
+    def enabled?
+      return false unless installed?
+
+      Yast::Service.Enabled?(SERVICE)
+    end
+
+    class << self
+      # Singleton instance
+      def instance
+        create_instance unless @instance
+        @instance
+      end
+
+      # Enforce a new clean instance
+      def create_instance
+        @instance = new
+      end
+
+      # Make sure only .instance and .create_instance can be used to
+      # create objects
+      private :new, :allocate
+    end
+  end
+end
